requirements.txt
streamlit==1.29.0
embedchain
langchain-text-splitters
pysqlite3-binary
chromadb
bcrypt
pymongo
python-docx
